Refactor out Ui class from Duke's reply functionality for clarity
